Assignment-management-System/requirements.txt
django-autoslug-iplweb
pillow
bootstrap4
